Trait libnmprofile::cli::ProfileCliDirective

pub trait ProfileCliDirective {
    fn run(self, action: ProfileAction);
}

Entrypoint for profile CLI directives.

Structs that define the CLI for a given profile are separate from the structs that implement the actual profile (see: WiFiProfileDirective and WiFiProfile in the wifi module). All CLI directives must implement this trait before integration into the main CLI.

Examples

For profiles that don’t accept CL arguments (see: wifi) run can simply be implemented as:

// some_profile.rs
use structopt::StructOpt;

use libnmprofile::cli::{ProfileAction, ProfileCliDirective};
use libnmprofile::profile::{self, Profile};

struct SomeProfile {}

impl Profile for SomeProfile {
    fn up(&self) {
        // some code...
    }

    fn down(&self) {
        // some code...
    }
}

#[derive(StructOpt)]
pub struct SomeProfileDirective {}

impl ProfileCliDirective for SomeProfileDirective {
    fn run(self, action: ProfileAction) {
        let profile = SomeProfile {};
        match action {
            ProfileAction::Up => profile.up(),
            ProfileAction::Down => profile.down(),
        }
    }
}

If a profile does accept CL arguments, they can be passed to the profile in the call to run:

// some_profile.rs
use structopt::StructOpt;

use libnmprofile::cli::{ProfileAction, ProfileCliDirective};
// snip...

struct SomeProfile {
    setting: String,
}

// snip...

#[derive(StructOpt)]
struct SomeProfileDirective {
    #[structopt(short, long)]
    setting: String,
}

impl ProfileCliDirective for SomeProfileDirective {
    fn run(self, action: ProfileAction) {
        let profile = SomeProfile {setting: self.setting};
        match action {
            ProfileAction::Up => profile.up(),
            ProfileAction::Down => profile.down(),
        }
    }
}

Required methods

fn run(self, action: ProfileAction)

Either apply or remove the profile.

Loading content...