Module libnmprofile::profile

Trait and utility functions for implementing profiles.

Traits

Profile
Network profile.

Functions

ethernet_is_connected
An ethernet cable is connected to the computer.

gen_ethernet_status
Status of the first ethernet device.

gen_wifi_status
Status of the first wifi device.

run_command
Run an arbitrary command using the provided shell, or Bash by default, and return stdout and stderr.

wait_for
Utility function to run a function continuously until it returns true.

wifi_is_available
The wireless card is enabled (status is not `unavailable).

wifi_is_connected
The wireless card is connected to a Wi-Fi network.