Function libnmprofile::profile::wait_for

pub fn wait_for<F>(predicate: F, sleep_for: Option<u64>) -> Result<()> where
    F: Fn() -> Result<bool>, 

Utility function to run a function continuously until it returns true.

Can be used with functions like wifi_is_connected to wait for a Wi-Fi connection before performing other steps in a profile.